Match native head output dtype

+++ scripts/evaluate_pytorch_w1a1.py
@@ -57,7 +57,7 @@ def variant_quantization(quant: dict, variant: dict) -> dict:
             "activation_scale": "mean_abs_f64_sum_f32_per_token",
             "integer_dot": "k_minus_2_popcount_xor",
             "scale_order": "dot_weight_activation_f32",
-            "output_dtype": "bfloat16",
+            "output_dtype": "float32",
         }
         for key, value in expected.items():
             if quant.get(key) != value:
